# Conflicts:
#	db/data_import_from_csv/careerpath.xlsx
#	db/data_import_from_csv/group.xlsx
#	db/data_import_from_csv/level.xlsx
#	db/data_import_from_csv/skill.xlsx
#	db/migrate/20200628063923_create_courses.rb
#	db/schema.rb
#	lib/tasks/import.rake

+++ db/migrate/20200629045415_create_articles.rb
@@ -4,7 +4,7 @@ class CreateArticles < ActiveRecord::Migration[6.0]
       t.string :title
       t.text :description
       t.time :duration
-      t.integer :view_number
+      t.bigint :view_number
       t.boolean :is_free
       t.text :link_file_attach
       t.timestamps
